# kathracatalogmanagerhelmservices/api_add_catalog_entry.go
/*
 * Kathra Catalog Manager
 *
 * KATHRA Catalog Management API permetting :   * Generate source's packages from templates   * Insert catalog entry from template      * Insert catalog entry from file     * Insert catalog entry from source repository    * Read catalog entries from catalog    * Read catalog details from catalog
 *
 * API version: 1.1.0-RC-SNAPSHOT
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package kathracatalogmanagerhelmservices

import (
	"net/http"
)

func AddCatalogEntryFromArchive(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusNotImplemented)
}

func AddCatalogEntryFromSourceRepositoryUrl(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusNotImplemented)
}

func AddCatalogEntryFromTemplate(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	AddCatalogEntryFromTemplateImpl(w, r)
}