# kathracatalogmanagerhelmservices/model_catalog_entry.go
/*
 * Kathra Catalog Manager
 *
 * KATHRA Catalog Management API permetting :   * Generate source's packages from templates   * Insert catalog entry from template      * Insert catalog entry from file     * Insert catalog entry from source repository    * Read catalog entries from catalog    * Read catalog details from catalog
 *
 * API version: 1.1.0-RC-SNAPSHOT
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package kathracatalogmanagerhelmservices

type CatalogEntry struct {

	// Name
	Name string `json:"name,omitempty"`

	// Description
	Description string `json:"description,omitempty"`

	// Description
	Documentation string `json:"documentation,omitempty"`

	// Version
	Version string `json:"version,omitempty"`

	// Uniq identifier
	ProviderId string `json:"providerId,omitempty"`

	// Catalog entry arguments
	Arguments []CatalogEntryArgument `json:"arguments,omitempty"`
}